Move UpdatePaletteEffects to Drawing.cpp
