+++ src/openrct2/drawing/Drawing.h
@@ -628,6 +628,7 @@ void MaskFn(
 std::optional<uint32_t> GetPaletteG1Index(colour_t paletteId);
 std::optional<PaletteMap> GetPaletteMapForColour(colour_t paletteId);
 void UpdatePalette(const uint8_t* colours, int32_t start_index, int32_t num_colours);
+void UpdatePaletteEffects();
 
 void RefreshVideo(bool recreateWindow);
 void ToggleWindowedMode();
